requirements.txt
falcon==1.4.1
gunicorn==19.9.0
numpy==1.16.2
toml==0.10.0
torchvision==0.2.2.post3